requirements.txt
albumentations>=1.3.0
gdown>=4.6.0
numpy>=1.24.2
onnxruntime-gpu>=1.13.1
onnxruntime>=1.13.1
opencv-python>=4.7.0.68
opencv-python-headless>=4.7.0.68
Pillow>=9.4.0
streamlit>=1.18.0
